feat: use catalog.json from @dcl/asset-packs
