+++ packages/@dcl/inspector/src/components/AssetsCatalog/types.ts
@@ -1,40 +1,4 @@
-import { ComponentName } from '@dcl/asset-packs'
-import { EditorComponentsTypes } from '../../lib/sdk/components'
-
-export interface Props {
-  catalog: AssetPack[]
-  error: Error | null
-  isLoading: boolean
-}
-
-// categories obtained from "builder.decentraland.org" catalog
-export const CATEGORIES = [
-  'decorations',
-  'structures',
-  'vehicles',
-  'furniture',
-  'appliances',
-  'ground',
-  'nature',
-  'tiles',
-  'year of the pig'
-].sort()
-
-export type AssetPack = {
-  id: string
-  name: string
-  thumbnail: string
-  assets: Asset[]
-}
-
-export type Asset = {
-  id: string
-  name: string
-  category: string
-  tags: string[]
-  contents: Record<string, string>
-  components: Partial<Record<ComponentName, EditorComponentsTypes['Actions'] | EditorComponentsTypes['Triggers']>>
-}
+import { Asset, AssetPack } from '../../lib/logic/catalog'
 
 export interface ThemeProps {
   onClick: (value: AssetPack) => void
@@ -48,4 +12,8 @@ export interface CategoriesProps {
 
 export interface AssetProps {
   value: Asset
+}
+
+export interface Props {
+  catalog: AssetPack[]
 }
