+++ packages/@dcl/inspector/src/components/ProjectAssetExplorer/types.ts
@@ -1,5 +1,4 @@
 import { ComponentName } from '@dcl/asset-packs'
-import { EditorComponentsTypes } from '../../lib/sdk/components'
 
 export interface IAsset {
   src: string
@@ -20,7 +19,7 @@ export interface AssetCellProp {
 export type AssetNodeBase = {
   name: string
   parent: AssetNodeFolder | null
-  components?: Partial<Record<ComponentName, EditorComponentsTypes['Actions'] | EditorComponentsTypes['Triggers']>>
+  components?: Partial<Record<ComponentName, any>>
 }
 
 export type AssetNodeItem = AssetNodeBase & {
